divided functions into different files
